move tweened and spring into separate modules
