+++ src/motion/index.js
@@ -1,0 +1,2 @@
+export * from './spring.js';
+export * from './tweened.js';
